chore: fixes from review feedback
