+++ apps/server/src/modules/provisioning/strategy/schulconnex/service/schulconnex-tool-provisioning.service.spec.ts
@@ -1,7 +1,7 @@
 import { createMock, type DeepMocked } from '@golevelup/ts-jest';
 import { Logger } from '@infra/logger';
 import { ObjectId } from '@mikro-orm/mongodb';
-import { SchoolSystemOptionsService, SchulConneXProvisioningOptions } from '@modules/legacy-school';
+import { SchoolSystemOptionsService } from '@modules/legacy-school';
 import { schoolSystemOptionsFactory } from '@modules/legacy-school/testing';
 import { MediaSourceDataFormat, mediaSourceFactory } from '@modules/media-source';
 import { ExternalToolMetadataUpdateService } from '@modules/media-source-sync';
@@ -9,18 +9,17 @@ import { MediumMetadataService } from '@modules/medium-metadata';
 import { mediumMetadataDtoFactory } from '@modules/medium-metadata/testing';
 import { MediaSchoolLicenseService } from '@modules/school-license';
 import { mediaSchoolLicenseFactory } from '@modules/school-license/testing';
-import { ExternalToolMedium, ExternalToolService, ExternalToolValidationService } from '@modules/tool';
+import { ExternalToolService, ExternalToolValidationService } from '@modules/tool';
+import { CustomParameterScope } from '@modules/tool/common/enum';
 import { ExternalTool } from '@modules/tool/external-tool/domain';
 import { ExternalToolMediumStatus } from '@modules/tool/external-tool/enum';
 import { customParameterFactory, externalToolFactory } from '@modules/tool/external-tool/testing';
 import { SchoolExternalToolService } from '@modules/tool/school-external-tool';
-import { SchoolExternalTool } from '@modules/tool/school-external-tool/domain';
 import { schoolExternalToolFactory } from '@modules/tool/school-external-tool/testing';
 import { type MediaUserLicense, MediaUserLicenseService } from '@modules/user-license';
 import { mediaUserLicenseFactory } from '@modules/user-license/testing';
 import { Test, type TestingModule } from '@nestjs/testing';
-import _ from 'lodash';
-import { ExternalToolMetadataUpdateFailedLoggable } from '../../../loggable';
+import { ExternalToolMetadataUpdateFailedLoggable, ExternalToolProvisioningFailedLoggable } from '../../../loggable';
 import { SchulconnexToolProvisioningService } from './schulconnex-tool-provisioning.service';
 
 describe(SchulconnexToolProvisioningService.name, () => {
@@ -41,42 +40,15 @@ describe(SchulconnexToolProvisioningService.name, () => {
 		module = await Test.createTestingModule({
 			providers: [
 				SchulconnexToolProvisioningService,
-				{
-					provide: ExternalToolService,
-					useValue: createMock<ExternalToolService>(),
-				},
-				{
-					provide: SchoolExternalToolService,
-					useValue: createMock<SchoolExternalToolService>(),
-				},
-				{
-					provide: MediaUserLicenseService,
-					useValue: createMock<MediaUserLicenseService>(),
-				},
-				{
-					provide: MediaSchoolLicenseService,
-					useValue: createMock<MediaSchoolLicenseService>(),
-				},
-				{
-					provide: SchoolSystemOptionsService,
-					useValue: createMock<SchoolSystemOptionsService>(),
-				},
-				{
-					provide: ExternalToolValidationService,
-					useValue: createMock<ExternalToolValidationService>(),
-				},
-				{
-					provide: MediumMetadataService,
-					useValue: createMock<MediumMetadataService>(),
-				},
-				{
-					provide: ExternalToolMetadataUpdateService,
-					useValue: createMock<ExternalToolMetadataUpdateService>(),
-				},
-				{
-					provide: Logger,
-					useValue: createMock<Logger>(),
-				},
+				{ provide: ExternalToolService, useValue: createMock<ExternalToolService>() },
+				{ provide: SchoolExternalToolService, useValue: createMock<SchoolExternalToolService>() },
+				{ provide: MediaUserLicenseService, useValue: createMock<MediaUserLicenseService>() },
+				{ provide: MediaSchoolLicenseService, useValue: createMock<MediaSchoolLicenseService>() },
+				{ provide: SchoolSystemOptionsService, useValue: createMock<SchoolSystemOptionsService>() },
+				{ provide: ExternalToolValidationService, useValue: createMock<ExternalToolValidationService>() },
+				{ provide: MediumMetadataService, useValue: createMock<MediumMetadataService>() },
+				{ provide: ExternalToolMetadataUpdateService, useValue: createMock<ExternalToolMetadataUpdateService>() },
+				{ provide: Logger, useValue: createMock<Logger>() },
 			],
 		}).compile();
 
@@ -100,733 +72,462 @@ describe(SchulconnexToolProvisioningService.name, () => {
 		jest.resetAllMocks();
 	});
 
-	describe('provisionSchoolExternalTool', () => {
-		describe('when schoolExternalTool has to be provisioned', () => {
-			const setup = () => {
-				const schoolId = new ObjectId().toHexString();
-				const userId = new ObjectId().toHexString();
-				const systemId = new ObjectId().toHexString();
-				const { provisioningOptions } = schoolSystemOptionsFactory.build({
-					provisioningOptions: { schoolExternalToolProvisioningEnabled: true },
-				});
-				const mediaSource = mediaSourceFactory.build({ sourceId: 'shared-source', format: undefined });
-				const mediaUserLicense: MediaUserLicense = mediaUserLicenseFactory.build({
-					userId,
-					mediumId: 'shared-medium',
-					mediaSource,
-				});
-				const mediaSchoolLicense = mediaSchoolLicenseFactory.build({
-					schoolId,
-					mediumId: mediaUserLicense.mediumId,
-					mediaSource,
-				});
-				const externalTool: ExternalTool = externalToolFactory
-					.withMedium({
-						mediumId: mediaUserLicense.mediumId,
-						mediaSourceId: mediaUserLicense.mediaSource?.sourceId,
-					})
-					.build();
-				const schoolExternalTool: SchoolExternalTool = schoolExternalToolFactory.build({
-					toolId: externalTool.id,
-					schoolId,
-				});
+	type Scenario = {
+		userId: string;
+		schoolId: string;
+		systemId: string;
+		license: MediaUserLicense;
+		externalTool: ExternalTool;
+	};
 
-				schoolSystemOptionsService.getProvisioningOptions.mockResolvedValueOnce(provisioningOptions);
-				mediaUserLicenseService.getMediaUserLicensesForUser.mockResolvedValueOnce([mediaUserLicense]);
-				mediaSchoolLicenseService.findMediaSchoolLicensesBySchoolId.mockResolvedValueOnce([mediaSchoolLicense]);
-				externalToolService.findExternalToolByMedium.mockResolvedValueOnce(externalTool);
-				schoolExternalToolService.findSchoolExternalTools.mockResolvedValueOnce([]);
-				schoolExternalToolService.findSchoolExternalTools.mockResolvedValueOnce([schoolExternalTool]);
+	const createActiveExternalTool = (license: MediaUserLicense): ExternalTool =>
+		externalToolFactory
+			.withMedium({
+				mediumId: license.mediumId,
+				mediaSourceId: license.mediaSource?.sourceId,
+				status: ExternalToolMediumStatus.ACTIVE,
+			})
+			.build({ parameters: [] });
 
-				return {
-					userId,
-					schoolId,
-					systemId,
-					mediaUserLicense,
-					externalTool,
-					schoolExternalTool,
-				};
-			};
-
-			it('should get provisioning options', async () => {
-				const { userId, schoolId, systemId } = setup();
-
-				await service.provisionSchoolExternalTools(userId, schoolId, systemId);
-
-				expect(schoolSystemOptionsService.getProvisioningOptions).toHaveBeenCalledWith(
-					SchulConneXProvisioningOptions,
-					schoolId,
-					systemId
-				);
-			});
-
-			it('should get media user licenses', async () => {
-				const { userId, schoolId, systemId } = setup();
-
-				await service.provisionSchoolExternalTools(userId, schoolId, systemId);
-
-				expect(mediaUserLicenseService.getMediaUserLicensesForUser).toHaveBeenCalledWith(userId);
-			});
-
-			it('should get external tool', async () => {
-				const { userId, schoolId, systemId, mediaUserLicense } = setup();
-
-				await service.provisionSchoolExternalTools(userId, schoolId, systemId);
-
-				expect(externalToolService.findExternalToolByMedium).toHaveBeenCalledWith(
-					mediaUserLicense.mediumId,
-					mediaUserLicense.mediaSource?.sourceId
-				);
-				expect(externalToolService.findExternalToolByMedium).toHaveBeenCalledTimes(1);
-			});
-
-			it('should look for a school external tool', async () => {
-				const { userId, schoolId, systemId, externalTool } = setup();
-
-				await service.provisionSchoolExternalTools(userId, schoolId, systemId);
-
-				expect(schoolExternalToolService.findSchoolExternalTools).toHaveBeenCalledWith({
-					schoolId,
-					toolId: externalTool.id,
-				});
-			});
-
-			it('should save a school external tool', async () => {
-				const { userId, schoolId, systemId, externalTool } = setup();
-
-				await service.provisionSchoolExternalTools(userId, schoolId, systemId);
-
-				expect(schoolExternalToolService.saveSchoolExternalTool).toHaveBeenCalledWith<[SchoolExternalTool]>(
-					new SchoolExternalTool({
-						id: expect.any(String),
-						toolId: externalTool.id,
-						isDeactivated: false,
-						schoolId,
-						parameters: [],
-					})
-				);
-			});
+	const createScenario = (overrides: Partial<Scenario> = {}): Scenario => {
+		const userId = new ObjectId().toHexString();
+		const schoolId = new ObjectId().toHexString();
+		const systemId = new ObjectId().toHexString();
+		const license = mediaUserLicenseFactory.build({ userId });
+		const externalTool = createActiveExternalTool(license);
+		const scenario: Scenario = { userId, schoolId, systemId, license, externalTool, ...overrides };
+		const { provisioningOptions } = schoolSystemOptionsFactory.build({
+			provisioningOptions: { schoolExternalToolProvisioningEnabled: true },
 		});
 
-		describe('when user and school licenses share a mediumId without a media source', () => {
-			it('should look up and provision the external tool only once', async () => {
-				const schoolId = new ObjectId().toHexString();
-				const userId = new ObjectId().toHexString();
-				const systemId = new ObjectId().toHexString();
-				const { provisioningOptions } = schoolSystemOptionsFactory.build({
-					provisioningOptions: { schoolExternalToolProvisioningEnabled: true },
-				});
-				const mediumId = 'shared-medium';
-				const mediaUserLicense = mediaUserLicenseFactory.build({ userId, mediumId, mediaSource: undefined });
-				const mediaSchoolLicense = mediaSchoolLicenseFactory.build({ schoolId, mediumId, mediaSource: undefined });
-				const externalTool = externalToolFactory.withMedium({ mediumId, mediaSourceId: undefined }).build();
+		schoolSystemOptionsService.getProvisioningOptions.mockResolvedValue(provisioningOptions);
+		mediaUserLicenseService.getMediaUserLicensesForUser.mockResolvedValue([scenario.license]);
+		mediaSchoolLicenseService.findMediaSchoolLicensesBySchoolId.mockResolvedValue([]);
+		externalToolService.findExternalToolByMedium.mockResolvedValue(scenario.externalTool);
+		externalToolService.createExternalTool.mockImplementation((tool) => Promise.resolve(tool));
+		externalToolValidationService.validateCreate.mockResolvedValue();
+		schoolExternalToolService.findSchoolExternalTools.mockResolvedValue([]);
+		schoolExternalToolService.saveSchoolExternalTool.mockImplementation((tool) => Promise.resolve(tool));
 
-				schoolSystemOptionsService.getProvisioningOptions.mockResolvedValueOnce(provisioningOptions);
-				mediaUserLicenseService.getMediaUserLicensesForUser.mockResolvedValueOnce([mediaUserLicense]);
-				mediaSchoolLicenseService.findMediaSchoolLicensesBySchoolId.mockResolvedValueOnce([mediaSchoolLicense]);
-				externalToolService.findExternalToolByMedium.mockResolvedValueOnce(externalTool);
-				schoolExternalToolService.findSchoolExternalTools.mockResolvedValueOnce([]);
+		return scenario;
+	};
 
-				await service.provisionSchoolExternalTools(userId, schoolId, systemId);
+	const provision = (scenario: Scenario): Promise<void> =>
+		service.provisionSchoolExternalTools(scenario.userId, scenario.schoolId, scenario.systemId);
 
-				expect(externalToolService.findExternalToolByMedium).toHaveBeenCalledWith(mediumId, undefined);
-				expect(externalToolService.findExternalToolByMedium).toHaveBeenCalledTimes(1);
-				expect(schoolExternalToolService.saveSchoolExternalTool).toHaveBeenCalledTimes(1);
-			});
-		});
-
-		describe('when schoolExternalToolProvisioningEnabled is false', () => {
-			const setup = () => {
-				const schoolId = new ObjectId().toHexString();
-				const userId = new ObjectId().toHexString();
-				const systemId = new ObjectId().toHexString();
+	describe('provisionSchoolExternalTools', () => {
+		describe('when provisioning is disabled', () => {
+			const createDisabledScenario = (): Scenario => {
+				const scenario = createScenario();
 				const { provisioningOptions } = schoolSystemOptionsFactory.build({});
-
-				schoolSystemOptionsService.getProvisioningOptions.mockResolvedValueOnce(provisioningOptions);
-
-				return {
-					userId,
-					schoolId,
-					systemId,
-				};
+				schoolSystemOptionsService.getProvisioningOptions.mockResolvedValue(provisioningOptions);
+				return scenario;
 			};
 
-			it('should not save a schoolExternalTool', async () => {
-				const { userId, schoolId, systemId } = setup();
+			it('should not load user licenses', async () => {
+				const scenario = createDisabledScenario();
 
-				await service.provisionSchoolExternalTools(userId, schoolId, systemId);
+				await provision(scenario);
 
-				expect(schoolExternalToolService.saveSchoolExternalTool).not.toHaveBeenCalled();
+				expect(mediaUserLicenseService.getMediaUserLicensesForUser).not.toHaveBeenCalled();
+			});
+
+			it('should not load school licenses', async () => {
+				const scenario = createDisabledScenario();
+
+				await provision(scenario);
+
+				expect(mediaSchoolLicenseService.findMediaSchoolLicensesBySchoolId).not.toHaveBeenCalled();
+			});
+
+			it('should not search for external tools', async () => {
+				const scenario = createDisabledScenario();
+
+				await provision(scenario);
+
+				expect(externalToolService.findExternalToolByMedium).not.toHaveBeenCalled();
 			});
 		});
 
-		describe('when no mediaUserLicenses were found', () => {
-			const setup = () => {
-				const schoolId = new ObjectId().toHexString();
-				const userId = new ObjectId().toHexString();
-				const systemId = new ObjectId().toHexString();
-				const { provisioningOptions } = schoolSystemOptionsFactory.build({
-					provisioningOptions: { schoolExternalToolProvisioningEnabled: true },
-				});
-				const externalTool: ExternalTool = externalToolFactory.build();
-
-				schoolSystemOptionsService.getProvisioningOptions.mockResolvedValueOnce(provisioningOptions);
-				mediaUserLicenseService.getMediaUserLicensesForUser.mockResolvedValueOnce([]);
-				mediaSchoolLicenseService.findMediaSchoolLicensesBySchoolId.mockResolvedValueOnce([]);
-				externalToolService.findExternalToolByMedium.mockResolvedValueOnce(externalTool);
-				schoolExternalToolService.findSchoolExternalTools.mockResolvedValueOnce([]);
-
-				return {
-					userId,
-					schoolId,
-					systemId,
-				};
+		describe('when no licenses are available', () => {
+			const createNoLicenseScenario = (): Scenario => {
+				const scenario = createScenario();
+				mediaUserLicenseService.getMediaUserLicensesForUser.mockResolvedValue([]);
+				return scenario;
 			};
 
-			it('should not save a schoolExternalTool', async () => {
-				const { userId, schoolId, systemId } = setup();
+			it('should not search for external tools', async () => {
+				const scenario = createNoLicenseScenario();
 
-				await service.provisionSchoolExternalTools(userId, schoolId, systemId);
+				await provision(scenario);
 
-				expect(schoolExternalToolService.saveSchoolExternalTool).not.toHaveBeenCalled();
+				expect(externalToolService.findExternalToolByMedium).not.toHaveBeenCalled();
 			});
-		});
-
-		describe('when no external tool was found and no template exists', () => {
-			const setup = () => {
-				const schoolId = new ObjectId().toHexString();
-				const userId = new ObjectId().toHexString();
-				const systemId = new ObjectId().toHexString();
-				const { provisioningOptions } = schoolSystemOptionsFactory.build({
-					provisioningOptions: { schoolExternalToolProvisioningEnabled: true },
-				});
-				const mediaUserLicenses: MediaUserLicense[] = [mediaUserLicenseFactory.build({ userId })];
-				const externalTool: ExternalTool = externalToolFactory
-					.withMedium({ status: ExternalToolMediumStatus.DRAFT })
-					.build();
-
-				schoolSystemOptionsService.getProvisioningOptions.mockResolvedValueOnce(provisioningOptions);
-				mediaUserLicenseService.getMediaUserLicensesForUser.mockResolvedValueOnce(mediaUserLicenses);
-				mediaSchoolLicenseService.findMediaSchoolLicensesBySchoolId.mockResolvedValueOnce([]);
-				externalToolService.findExternalToolByMedium.mockResolvedValueOnce(externalTool);
-				schoolExternalToolService.findSchoolExternalTools.mockResolvedValueOnce([]);
-
-				return {
-					userId,
-					schoolId,
-					systemId,
-				};
-			};
 
 			it('should not create a school external tool', async () => {
-				const { userId, schoolId, systemId } = setup();
+				const scenario = createNoLicenseScenario();
 
-				await service.provisionSchoolExternalTools(userId, schoolId, systemId);
-
-				expect(schoolExternalToolService.saveSchoolExternalTool).not.toHaveBeenCalled();
-			});
-		});
-
-		describe('when no external tool was found, but a template for a media source without format', () => {
-			const setup = () => {
-				const schoolId = new ObjectId().toHexString();
-				const userId = new ObjectId().toHexString();
-				const systemId = new ObjectId().toHexString();
-				const { provisioningOptions } = schoolSystemOptionsFactory.build({
-					provisioningOptions: { schoolExternalToolProvisioningEnabled: true },
-				});
-				const mediaSource = mediaSourceFactory.build({ sourceId: 'mediaSourceId', format: undefined });
-				const mediaUserLicense = mediaUserLicenseFactory.build({
-					userId,
-					mediaSource,
-					mediumId: 'mediumId',
-				});
-				const initialExternalToolTemplate = externalToolFactory
-					.withFileRecordRef()
-					.withMedium({
-						mediaSourceId: mediaUserLicense.mediaSource?.sourceId,
-						status: ExternalToolMediumStatus.TEMPLATE,
-					})
-					.build();
-				const externalTool = _.cloneDeep(initialExternalToolTemplate);
-
-				schoolSystemOptionsService.getProvisioningOptions.mockResolvedValueOnce(provisioningOptions);
-				mediaUserLicenseService.getMediaUserLicensesForUser.mockResolvedValueOnce([mediaUserLicense]);
-				mediaSchoolLicenseService.findMediaSchoolLicensesBySchoolId.mockResolvedValueOnce([]);
-				externalToolService.findExternalToolByMedium.mockResolvedValueOnce(null);
-				externalToolService.findTemplate.mockResolvedValueOnce(externalTool);
-				externalToolService.createExternalTool.mockResolvedValueOnce(externalTool);
-				schoolExternalToolService.findSchoolExternalTools.mockResolvedValueOnce([]);
-
-				return {
-					userId,
-					schoolId,
-					systemId,
-					mediaUserLicense,
-					initialExternalToolTemplate,
-					externalTool,
-					mediaSource,
-				};
-			};
-
-			it('should save an external tool draft', async () => {
-				const { userId, schoolId, systemId, initialExternalToolTemplate, mediaUserLicense, mediaSource } = setup();
-
-				await service.provisionSchoolExternalTools(userId, schoolId, systemId);
-
-				expect(externalToolService.createExternalTool).toHaveBeenCalledWith<[ExternalTool]>(
-					new ExternalTool({
-						...initialExternalToolTemplate.getProps(),
-						id: expect.any(String),
-						name: `Draft: ${mediaSource.sourceId} ${mediaUserLicense.mediumId}`,
-						thumbnail: undefined,
-						medium: new ExternalToolMedium({
-							mediumId: mediaUserLicense.mediumId,
-							mediaSourceId: mediaSource.sourceId,
-							status: ExternalToolMediumStatus.DRAFT,
-							publisher: initialExternalToolTemplate.medium?.publisher,
-						}),
-					})
-				);
-			});
-
-			it('should not load metadata', async () => {
-				const { userId, schoolId, systemId } = setup();
-
-				await service.provisionSchoolExternalTools(userId, schoolId, systemId);
-
-				expect(mediumMetadataService.getMetadataItem).not.toHaveBeenCalled();
-			});
-
-			it('should not save a schoolExternalTool', async () => {
-				const { userId, schoolId, systemId } = setup();
-
-				await service.provisionSchoolExternalTools(userId, schoolId, systemId);
+				await provision(scenario);
 
 				expect(schoolExternalToolService.saveSchoolExternalTool).not.toHaveBeenCalled();
 			});
 		});
 
-		describe('when no external tool was found, but a template for a media source with format', () => {
-			const setup = () => {
-				const schoolId = new ObjectId().toHexString();
-				const userId = new ObjectId().toHexString();
-				const systemId = new ObjectId().toHexString();
-				const updatedToolName = 'New Tool';
-				const { provisioningOptions } = schoolSystemOptionsFactory.build({
-					provisioningOptions: { schoolExternalToolProvisioningEnabled: true },
-				});
-				const mediaSource = mediaSourceFactory.build({
-					sourceId: 'mediaSourceId',
-					format: MediaSourceDataFormat.VIDIS,
-				});
-				const mediaUserLicense = mediaUserLicenseFactory.build({
-					userId,
-					mediaSource,
-					mediumId: 'mediumId',
-				});
-				const initialExternalToolTemplate = externalToolFactory
-					.withFileRecordRef()
-					.withMedium({
-						mediaSourceId: mediaUserLicense.mediaSource?.sourceId,
-						status: ExternalToolMediumStatus.TEMPLATE,
-					})
-					.build();
-				const externalTool = _.cloneDeep(initialExternalToolTemplate);
-				const mediumMetadata = mediumMetadataDtoFactory.build({
-					name: updatedToolName,
-				});
+		describe('when an active external tool exists', () => {
+			it('should create a school external tool', async () => {
+				const scenario = createScenario();
 
-				schoolSystemOptionsService.getProvisioningOptions.mockResolvedValueOnce(provisioningOptions);
-				mediaUserLicenseService.getMediaUserLicensesForUser.mockResolvedValueOnce([mediaUserLicense]);
-				mediaSchoolLicenseService.findMediaSchoolLicensesBySchoolId.mockResolvedValueOnce([]);
-				externalToolService.findExternalToolByMedium.mockResolvedValueOnce(null);
-				externalToolService.findTemplate.mockResolvedValueOnce(externalTool);
-				externalToolService.createExternalTool.mockResolvedValueOnce(externalTool);
-				mediumMetadataService.getMetadataItem.mockResolvedValueOnce(mediumMetadata);
-				externalToolMetadataUpdateService.updateExternalToolWithMetadata.mockImplementation(
-					async (externalTool: ExternalTool) => {
-						externalTool.name = updatedToolName;
-						await Promise.resolve();
-					}
-				);
-				schoolExternalToolService.findSchoolExternalTools.mockResolvedValueOnce([]);
+				await provision(scenario);
 
-				return {
-					userId,
-					schoolId,
-					systemId,
-					mediaUserLicense,
-					initialExternalToolTemplate,
-					externalTool,
-					updatedToolName,
-					mediaSource,
-				};
-			};
-
-			it('should save an active external tool', async () => {
-				const {
-					userId,
-					schoolId,
-					systemId,
-					initialExternalToolTemplate,
-					mediaUserLicense,
-					mediaSource,
-					updatedToolName,
-				} = setup();
-
-				await service.provisionSchoolExternalTools(userId, schoolId, systemId);
-
-				expect(externalToolService.createExternalTool).toHaveBeenCalledWith<[ExternalTool]>(
-					new ExternalTool({
-						...initialExternalToolTemplate.getProps(),
-						id: expect.any(String),
-						name: updatedToolName,
-						thumbnail: undefined,
-						medium: new ExternalToolMedium({
-							mediumId: mediaUserLicense.mediumId,
-							mediaSourceId: mediaSource.sourceId,
-							status: ExternalToolMediumStatus.ACTIVE,
-							publisher: initialExternalToolTemplate.medium?.publisher,
-						}),
-					})
-				);
-			});
-
-			it('should save a school external tool', async () => {
-				const { userId, schoolId, systemId, externalTool } = setup();
-
-				await service.provisionSchoolExternalTools(userId, schoolId, systemId);
-
-				expect(schoolExternalToolService.saveSchoolExternalTool).toHaveBeenCalledWith<[SchoolExternalTool]>(
-					new SchoolExternalTool({
-						id: expect.any(String),
-						toolId: externalTool.id,
+				expect(schoolExternalToolService.saveSchoolExternalTool).toHaveBeenCalledWith(
+					expect.objectContaining({
+						toolId: scenario.externalTool.id,
+						schoolId: scenario.schoolId,
 						isDeactivated: false,
-						schoolId,
 						parameters: [],
 					})
 				);
 			});
-		});
 
-		describe('when loading the metadata for an external tool fails', () => {
-			const setup = () => {
-				const schoolId = new ObjectId().toHexString();
-				const userId = new ObjectId().toHexString();
-				const systemId = new ObjectId().toHexString();
-				const { provisioningOptions } = schoolSystemOptionsFactory.build({
-					provisioningOptions: { schoolExternalToolProvisioningEnabled: true },
-				});
-				const mediaSource = mediaSourceFactory.build({
-					sourceId: 'mediaSourceId',
-					format: MediaSourceDataFormat.VIDIS,
-				});
-				const mediaUserLicense = mediaUserLicenseFactory.build({
-					userId,
-					mediaSource,
-					mediumId: 'mediumId',
-				});
-				const initialExternalToolTemplate = externalToolFactory
-					.withFileRecordRef()
-					.withMedium({
-						mediaSourceId: mediaUserLicense.mediaSource?.sourceId,
-						status: ExternalToolMediumStatus.TEMPLATE,
-					})
-					.build();
-				const externalTool = _.cloneDeep(initialExternalToolTemplate);
-				const error = new Error();
-
-				schoolSystemOptionsService.getProvisioningOptions.mockResolvedValueOnce(provisioningOptions);
-				mediaUserLicenseService.getMediaUserLicensesForUser.mockResolvedValueOnce([mediaUserLicense]);
-				mediaSchoolLicenseService.findMediaSchoolLicensesBySchoolId.mockResolvedValueOnce([]);
-				externalToolService.findExternalToolByMedium.mockResolvedValueOnce(null);
-				externalToolService.findTemplate.mockResolvedValueOnce(externalTool);
-				externalToolService.createExternalTool.mockResolvedValueOnce(externalTool);
-				mediumMetadataService.getMetadataItem.mockRejectedValueOnce(error);
-				schoolExternalToolService.findSchoolExternalTools.mockResolvedValueOnce([]);
-
-				return {
-					userId,
-					schoolId,
-					systemId,
-					mediaUserLicense,
-					initialExternalToolTemplate,
-					externalTool,
-					mediaSource,
-					error,
-				};
-			};
-
-			it('should log a warning', async () => {
-				const { userId, schoolId, systemId, mediaUserLicense, error } = setup();
-
-				await service.provisionSchoolExternalTools(userId, schoolId, systemId);
-
-				expect(logger.warning).toHaveBeenCalledWith(
-					new ExternalToolMetadataUpdateFailedLoggable(
-						expect.any(ExternalTool) as unknown as ExternalTool,
-						mediaUserLicense,
-						error
-					)
-				);
-			});
-
-			it('should save an external tool draft', async () => {
-				const { userId, schoolId, systemId, initialExternalToolTemplate, mediaUserLicense, mediaSource } = setup();
-
-				await service.provisionSchoolExternalTools(userId, schoolId, systemId);
-
-				expect(externalToolService.createExternalTool).toHaveBeenCalledWith<[ExternalTool]>(
-					new ExternalTool({
-						...initialExternalToolTemplate.getProps(),
-						id: expect.any(String),
-						name: `Draft: ${mediaSource.sourceId} ${mediaUserLicense.mediumId}`,
-						thumbnail: undefined,
-						medium: new ExternalToolMedium({
-							mediumId: mediaUserLicense.mediumId,
-							mediaSourceId: mediaSource.sourceId,
-							status: ExternalToolMediumStatus.DRAFT,
-							publisher: initialExternalToolTemplate.medium?.publisher,
-						}),
-					})
-				);
-			});
-
-			it('should not save a schoolExternalTool', async () => {
-				const { userId, schoolId, systemId } = setup();
-
-				await service.provisionSchoolExternalTools(userId, schoolId, systemId);
-
-				expect(schoolExternalToolService.saveSchoolExternalTool).not.toHaveBeenCalled();
-			});
-		});
-
-		describe('when the validation for creating an external tool fails', () => {
-			const setup = () => {
-				const schoolId = new ObjectId().toHexString();
-				const userId = new ObjectId().toHexString();
-				const systemId = new ObjectId().toHexString();
-				const { provisioningOptions } = schoolSystemOptionsFactory.build({
-					provisioningOptions: { schoolExternalToolProvisioningEnabled: true },
-				});
-				const mediaSource = mediaSourceFactory.build({ sourceId: 'mediaSourceId', format: undefined });
-				const mediaUserLicense = mediaUserLicenseFactory.build({
-					userId,
-					mediaSource,
-					mediumId: 'mediumId',
-				});
-				const initialExternalToolTemplate = externalToolFactory
-					.withFileRecordRef()
-					.withMedium({
-						mediaSourceId: mediaUserLicense.mediaSource?.sourceId,
-						status: ExternalToolMediumStatus.TEMPLATE,
-					})
-					.build();
-				const externalTool = _.cloneDeep(initialExternalToolTemplate);
-
-				schoolSystemOptionsService.getProvisioningOptions.mockResolvedValueOnce(provisioningOptions);
-				mediaUserLicenseService.getMediaUserLicensesForUser.mockResolvedValueOnce([mediaUserLicense]);
-				mediaSchoolLicenseService.findMediaSchoolLicensesBySchoolId.mockResolvedValueOnce([]);
-				externalToolService.findExternalToolByMedium.mockResolvedValueOnce(null);
-				externalToolService.findTemplate.mockResolvedValueOnce(externalTool);
-				externalToolValidationService.validateCreate.mockRejectedValueOnce(new Error());
-				schoolExternalToolService.findSchoolExternalTools.mockResolvedValueOnce([]);
-
-				return {
-					userId,
-					schoolId,
-					systemId,
-					mediaUserLicense,
-					initialExternalToolTemplate,
-					externalTool,
-					mediaSource,
-				};
-			};
-
-			it('should not save an external tool', async () => {
-				const { userId, schoolId, systemId } = setup();
-
-				await service.provisionSchoolExternalTools(userId, schoolId, systemId);
-
-				expect(externalToolService.createExternalTool).not.toHaveBeenCalled();
-			});
-
-			it('should not save a school external tool', async () => {
-				const { userId, schoolId, systemId } = setup();
-
-				await service.provisionSchoolExternalTools(userId, schoolId, systemId);
-
-				expect(schoolExternalToolService.saveSchoolExternalTool).not.toHaveBeenCalled();
-			});
-		});
-
-		describe('when the external tool template has no medium', () => {
-			const setup = () => {
-				const schoolId = new ObjectId().toHexString();
-				const userId = new ObjectId().toHexString();
-				const systemId = new ObjectId().toHexString();
-				const { provisioningOptions } = schoolSystemOptionsFactory.build({
-					provisioningOptions: { schoolExternalToolProvisioningEnabled: true },
-				});
-				const mediaSource = mediaSourceFactory.build({ sourceId: 'mediaSourceId', format: undefined });
-				const mediaUserLicense = mediaUserLicenseFactory.build({
-					userId,
-					mediaSource,
-					mediumId: 'mediumId',
-				});
-				const externalToolTemplate = externalToolFactory.build({
-					medium: undefined,
-				});
-
-				schoolSystemOptionsService.getProvisioningOptions.mockResolvedValueOnce(provisioningOptions);
-				mediaUserLicenseService.getMediaUserLicensesForUser.mockResolvedValueOnce([mediaUserLicense]);
-				mediaSchoolLicenseService.findMediaSchoolLicensesBySchoolId.mockResolvedValueOnce([]);
-				externalToolService.findExternalToolByMedium.mockResolvedValueOnce(null);
-				externalToolService.findTemplate.mockResolvedValueOnce(externalToolTemplate);
-				schoolExternalToolService.findSchoolExternalTools.mockResolvedValueOnce([]);
-
-				return {
-					userId,
-					schoolId,
-					systemId,
-					mediaUserLicense,
-					mediaSource,
-				};
-			};
-
-			it('should not save an external tool', async () => {
-				const { userId, schoolId, systemId } = setup();
-
-				await service.provisionSchoolExternalTools(userId, schoolId, systemId);
-
-				expect(externalToolService.createExternalTool).not.toHaveBeenCalled();
-			});
-
-			it('should save save a school external tool', async () => {
-				const { userId, schoolId, systemId } = setup();
-
-				await service.provisionSchoolExternalTools(userId, schoolId, systemId);
-
-				expect(schoolExternalToolService.saveSchoolExternalTool).not.toHaveBeenCalled();
-			});
-		});
-
-		describe('when an external tool was found, but is only a draft', () => {
-			const setup = () => {
-				const schoolId = new ObjectId().toHexString();
-				const userId = new ObjectId().toHexString();
-				const systemId = new ObjectId().toHexString();
-				const { provisioningOptions } = schoolSystemOptionsFactory.build({
-					provisioningOptions: { schoolExternalToolProvisioningEnabled: true },
-				});
-				const mediaUserLicenses: MediaUserLicense[] = [mediaUserLicenseFactory.build({ userId })];
-
-				schoolSystemOptionsService.getProvisioningOptions.mockResolvedValueOnce(provisioningOptions);
-				mediaUserLicenseService.getMediaUserLicensesForUser.mockResolvedValueOnce(mediaUserLicenses);
-				mediaSchoolLicenseService.findMediaSchoolLicensesBySchoolId.mockResolvedValueOnce([]);
-				externalToolService.findExternalToolByMedium.mockResolvedValueOnce(null);
-				externalToolService.findTemplate.mockResolvedValueOnce(null);
-				schoolExternalToolService.findSchoolExternalTools.mockResolvedValueOnce([]);
-
-				return {
-					userId,
-					schoolId,
-					systemId,
-				};
-			};
-
-			it('should not create a school external tool', async () => {
-				const { userId, schoolId, systemId } = setup();
-
-				await service.provisionSchoolExternalTools(userId, schoolId, systemId);
-
-				expect(schoolExternalToolService.saveSchoolExternalTool).not.toHaveBeenCalled();
-			});
-		});
-
-		describe('when tool has non-global parameters', () => {
-			const setup = () => {
-				const schoolId = new ObjectId().toHexString();
-				const userId = new ObjectId().toHexString();
-				const systemId = new ObjectId().toHexString();
-				const { provisioningOptions } = schoolSystemOptionsFactory.build({
-					provisioningOptions: { schoolExternalToolProvisioningEnabled: true },
-				});
-				const mediaUserLicenses: MediaUserLicense[] = [mediaUserLicenseFactory.build({ userId })];
-				const externalTool: ExternalTool = externalToolFactory
-					.withMedium({
-						mediumId: mediaUserLicenses[0].mediumId,
-						mediaSourceId: mediaUserLicenses[0].mediaSource?.sourceId,
-					})
-					.build({
-						parameters: [customParameterFactory.build()],
+			describe('when a school external tool already exists', () => {
+				it('should not create a duplicate school external tool', async () => {
+					const scenario = createScenario();
+					const existingSchoolExternalTool = schoolExternalToolFactory.build({
+						schoolId: scenario.schoolId,
+						toolId: scenario.externalTool.id,
 					});
+					schoolExternalToolService.findSchoolExternalTools.mockResolvedValue([existingSchoolExternalTool]);
 
-				schoolSystemOptionsService.getProvisioningOptions.mockResolvedValueOnce(provisioningOptions);
-				mediaUserLicenseService.getMediaUserLicensesForUser.mockResolvedValueOnce(mediaUserLicenses);
-				mediaSchoolLicenseService.findMediaSchoolLicensesBySchoolId.mockResolvedValueOnce([]);
-				externalToolService.findExternalToolByMedium.mockResolvedValueOnce(externalTool);
-				schoolExternalToolService.findSchoolExternalTools.mockResolvedValueOnce([]);
+					await provision(scenario);
 
-				return {
-					userId,
-					schoolId,
-					systemId,
+					expect(schoolExternalToolService.saveSchoolExternalTool).not.toHaveBeenCalled();
+				});
+			});
+
+			describe('when an inactive external tool exists', () => {
+				const createInactiveToolScenario = (): Scenario => {
+					const scenario = createScenario();
+					if (scenario.externalTool.medium) {
+						scenario.externalTool.medium.status = ExternalToolMediumStatus.DRAFT;
+					}
+					externalToolService.findExternalToolByMedium.mockResolvedValue(scenario.externalTool);
+					return scenario;
 				};
-			};
 
-			it('should not save a schoolExternalTool', async () => {
-				const { userId, schoolId, systemId } = setup();
+				it('should not save a school external tool', async () => {
+					const scenario = createInactiveToolScenario();
 
-				await service.provisionSchoolExternalTools(userId, schoolId, systemId);
+					await provision(scenario);
 
-				expect(schoolExternalToolService.saveSchoolExternalTool).not.toHaveBeenCalled();
+					expect(schoolExternalToolService.saveSchoolExternalTool).not.toHaveBeenCalled();
+				});
+
+				it('should log a provisioning failure', async () => {
+					const scenario = createInactiveToolScenario();
+
+					await provision(scenario);
+
+					expect(logger.warning).toHaveBeenCalledWith(expect.any(ExternalToolProvisioningFailedLoggable));
+				});
+			});
+
+			describe('when the tool has non-global parameters', () => {
+				const createNonGlobalParameterScenario = (): Scenario => {
+					const scenario = createScenario({
+						externalTool: externalToolFactory
+							.withMedium({
+								mediumId: 'medium-id',
+								mediaSourceId: 'media-source-id',
+								status: ExternalToolMediumStatus.ACTIVE,
+							})
+							.build({ parameters: [customParameterFactory.build({ scope: CustomParameterScope.SCHOOL })] }),
+					});
+					externalToolService.findExternalToolByMedium.mockResolvedValue(scenario.externalTool);
+					return scenario;
+				};
+
+				it('should not save a school external tool', async () => {
+					const scenario = createNonGlobalParameterScenario();
+
+					await provision(scenario);
+
+					expect(schoolExternalToolService.saveSchoolExternalTool).not.toHaveBeenCalled();
+				});
+
+				it('should log a provisioning failure', async () => {
+					const scenario = createNonGlobalParameterScenario();
+
+					await provision(scenario);
+
+					expect(logger.warning).toHaveBeenCalledWith(expect.any(ExternalToolProvisioningFailedLoggable));
+				});
 			});
 		});
 
-		describe('when schoolExternalTool already exists', () => {
-			const setup = () => {
-				const schoolId = new ObjectId().toHexString();
-				const userId = new ObjectId().toHexString();
-				const systemId = new ObjectId().toHexString();
-				const { provisioningOptions } = schoolSystemOptionsFactory.build({
-					provisioningOptions: { schoolExternalToolProvisioningEnabled: true },
+		describe('when creating an external tool from a template', () => {
+			const createTemplateScenario = (format?: MediaSourceDataFormat) => {
+				const mediaSource = mediaSourceFactory.build({ sourceId: 'media-source-id', format });
+				const license = mediaUserLicenseFactory.build({
+					mediaSource,
+					mediumId: 'medium-id',
 				});
-				const mediaUserLicenses: MediaUserLicense[] = [mediaUserLicenseFactory.build({ userId })];
-				const externalTool: ExternalTool = externalToolFactory
-					.withMedium({
-						mediumId: mediaUserLicenses[0].mediumId,
-						mediaSourceId: mediaUserLicenses[0].mediaSource?.sourceId,
-					})
+				const template = externalToolFactory
+					.withFileRecordRef()
+					.withMedium({ mediaSourceId: mediaSource.sourceId, status: ExternalToolMediumStatus.TEMPLATE })
 					.build();
-				const schoolExternalTool: SchoolExternalTool = schoolExternalToolFactory.build({
-					toolId: externalTool.id,
-					schoolId,
-				});
+				const scenario = createScenario({ license, externalTool: template });
 
-				schoolSystemOptionsService.getProvisioningOptions.mockResolvedValueOnce(provisioningOptions);
-				mediaUserLicenseService.getMediaUserLicensesForUser.mockResolvedValueOnce(mediaUserLicenses);
-				mediaSchoolLicenseService.findMediaSchoolLicensesBySchoolId.mockResolvedValueOnce([]);
-				externalToolService.findExternalToolByMedium.mockResolvedValueOnce(externalTool);
-				schoolExternalToolService.findSchoolExternalTools.mockResolvedValueOnce([schoolExternalTool]);
-
-				return {
-					userId,
-					schoolId,
-					systemId,
-					mediaUserLicenses,
-					externalTool,
-					schoolExternalTool,
-				};
+				externalToolService.findExternalToolByMedium.mockResolvedValue(null);
+				externalToolService.findTemplate.mockResolvedValue(template);
+				return { scenario, template, mediaSource };
 			};
 
-			it('should not save a schoolExternalTool', async () => {
-				const { userId, schoolId, systemId } = setup();
+			describe('when metadata is available', () => {
+				const createMetadataAvailableScenario = () => {
+					const { scenario, mediaSource } = createTemplateScenario(MediaSourceDataFormat.VIDIS);
+					const metadata = mediumMetadataDtoFactory.build({ name: 'Updated tool name' });
+					mediumMetadataService.getMetadataItem.mockResolvedValue(metadata);
+					externalToolMetadataUpdateService.updateExternalToolWithMetadata.mockImplementation((tool) => {
+						tool.name = metadata.name;
+						return Promise.resolve();
+					});
+					return { scenario, mediaSource, metadata };
+				};
 
-				await service.provisionSchoolExternalTools(userId, schoolId, systemId);
+				it('should create an active external tool', async () => {
+					const { scenario, mediaSource, metadata } = createMetadataAvailableScenario();
 
-				expect(schoolExternalToolService.saveSchoolExternalTool).not.toHaveBeenCalled();
+					await provision(scenario);
+
+					const [createdTool] = externalToolService.createExternalTool.mock.calls[0];
+					expect(createdTool).toEqual(
+						expect.objectContaining({
+							name: metadata.name,
+							medium: expect.objectContaining({
+								mediumId: scenario.license.mediumId,
+								mediaSourceId: mediaSource.sourceId,
+								status: ExternalToolMediumStatus.ACTIVE,
+							}),
+						})
+					);
+				});
+
+				it('should create a school external tool', async () => {
+					const { scenario } = createMetadataAvailableScenario();
+
+					await provision(scenario);
+
+					expect(schoolExternalToolService.saveSchoolExternalTool).toHaveBeenCalled();
+				});
+			});
+
+			describe('when metadata loading fails', () => {
+				const createMetadataFailureScenario = () => {
+					const { scenario, template } = createTemplateScenario(MediaSourceDataFormat.VIDIS);
+					const error = new Error('metadata unavailable');
+					mediumMetadataService.getMetadataItem.mockRejectedValue(error);
+					return { scenario, template, error };
+				};
+
+				it('should create a draft external tool', async () => {
+					const { scenario } = createMetadataFailureScenario();
+					const { mediaSource } = scenario.license;
+					const mediaSourceId = mediaSource?.sourceId ?? '-';
+
+					await provision(scenario);
+
+					const [createdTool] = externalToolService.createExternalTool.mock.calls[0];
+					expect(createdTool).toEqual(
+						expect.objectContaining({
+							name: `Draft: ${mediaSourceId} ${scenario.license.mediumId}`,
+							medium: expect.objectContaining({ status: ExternalToolMediumStatus.DRAFT }),
+						})
+					);
+				});
+
+				it('should log a metadata update warning', async () => {
+					const { scenario, error } = createMetadataFailureScenario();
+
+					await provision(scenario);
+
+					expect(logger.warning).toHaveBeenCalledWith(
+						new ExternalToolMetadataUpdateFailedLoggable(
+							expect.any(ExternalTool) as unknown as ExternalTool,
+							scenario.license,
+							error
+						)
+					);
+				});
+
+				it('should not create a school external tool', async () => {
+					const { scenario } = createMetadataFailureScenario();
+
+					await provision(scenario);
+
+					expect(schoolExternalToolService.saveSchoolExternalTool).not.toHaveBeenCalled();
+				});
+
+				it('should keep the created tool in draft status', async () => {
+					const { scenario, template } = createMetadataFailureScenario();
+
+					await provision(scenario);
+
+					expect(template.medium?.status).toBe(ExternalToolMediumStatus.DRAFT);
+				});
+			});
+
+			describe('when the template is missing', () => {
+				const createMissingTemplateScenario = (): Scenario => {
+					const { scenario } = createTemplateScenario();
+					externalToolService.findTemplate.mockResolvedValue(null);
+					return scenario;
+				};
+
+				it('should not create an external tool', async () => {
+					const scenario = createMissingTemplateScenario();
+
+					await provision(scenario);
+
+					expect(externalToolService.createExternalTool).not.toHaveBeenCalled();
+				});
+
+				it('should not create a school external tool', async () => {
+					const scenario = createMissingTemplateScenario();
+
+					await provision(scenario);
+
+					expect(schoolExternalToolService.saveSchoolExternalTool).not.toHaveBeenCalled();
+				});
+
+				it('should log a provisioning failure', async () => {
+					const scenario = createMissingTemplateScenario();
+
+					await provision(scenario);
+
+					expect(logger.warning).toHaveBeenCalledWith(expect.any(ExternalToolProvisioningFailedLoggable));
+				});
+			});
+
+			describe('when the template has no medium', () => {
+				const createTemplateWithoutMediumScenario = (): Scenario => {
+					const { scenario } = createTemplateScenario();
+					externalToolService.findTemplate.mockResolvedValue(externalToolFactory.build({ medium: undefined }));
+					return scenario;
+				};
+
+				it('should not create an external tool', async () => {
+					const scenario = createTemplateWithoutMediumScenario();
+
+					await provision(scenario);
+
+					expect(externalToolService.createExternalTool).not.toHaveBeenCalled();
+				});
+
+				it('should not create a school external tool', async () => {
+					const scenario = createTemplateWithoutMediumScenario();
+
+					await provision(scenario);
+
+					expect(schoolExternalToolService.saveSchoolExternalTool).not.toHaveBeenCalled();
+				});
+			});
+
+			describe('when validation fails', () => {
+				const createValidationFailureScenario = (): Scenario => {
+					const { scenario } = createTemplateScenario();
+					externalToolValidationService.validateCreate.mockRejectedValue(new Error('invalid tool'));
+					return scenario;
+				};
+
+				it('should not create an external tool', async () => {
+					const scenario = createValidationFailureScenario();
+
+					await provision(scenario);
+
+					expect(externalToolService.createExternalTool).not.toHaveBeenCalled();
+				});
+
+				it('should not create a school external tool', async () => {
+					const scenario = createValidationFailureScenario();
+
+					await provision(scenario);
+
+					expect(schoolExternalToolService.saveSchoolExternalTool).not.toHaveBeenCalled();
+				});
+
+				it('should log a provisioning failure', async () => {
+					const scenario = createValidationFailureScenario();
+
+					await provision(scenario);
+
+					expect(logger.warning).toHaveBeenCalledWith(expect.any(ExternalToolProvisioningFailedLoggable));
+				});
+			});
+		});
+
+		describe('processing multiple licenses', () => {
+			it('should provision different media with the same name', async () => {
+				const scenario = createScenario();
+				const secondLicense = mediaSchoolLicenseFactory.build({
+					schoolId: scenario.schoolId,
+					mediumId: 'second-medium-id',
+					mediaSource: mediaSourceFactory.build({ sourceId: 'second-media-source-id' }),
+				});
+				const secondTool = createActiveExternalTool(secondLicense as unknown as MediaUserLicense);
+				scenario.externalTool.name = 'same medium name';
+				secondTool.name = scenario.externalTool.name;
+				mediaSchoolLicenseService.findMediaSchoolLicensesBySchoolId.mockResolvedValue([secondLicense]);
+				externalToolService.findExternalToolByMedium.mockImplementation((mediumId) => {
+					if (mediumId === scenario.license.mediumId) {
+						return Promise.resolve(scenario.externalTool);
+					}
+					return Promise.resolve(secondTool);
+				});
+
+				await provision(scenario);
+
+				expect(schoolExternalToolService.saveSchoolExternalTool).toHaveBeenCalledTimes(2);
+				expect(schoolExternalToolService.saveSchoolExternalTool).toHaveBeenCalledWith(
+					expect.objectContaining({ toolId: scenario.externalTool.id, schoolId: scenario.schoolId })
+				);
+				expect(schoolExternalToolService.saveSchoolExternalTool).toHaveBeenCalledWith(
+					expect.objectContaining({ toolId: secondTool.id, schoolId: scenario.schoolId })
+				);
+			});
+
+			describe('when one license fails', () => {
+				const createPartialFailureScenario = (): { scenario: Scenario; secondTool: ExternalTool } => {
+					const scenario = createScenario();
+					const secondLicense = mediaSchoolLicenseFactory.build();
+					const secondTool = createActiveExternalTool(secondLicense as unknown as MediaUserLicense);
+					mediaSchoolLicenseService.findMediaSchoolLicensesBySchoolId.mockResolvedValue([secondLicense]);
+					externalToolService.findExternalToolByMedium.mockImplementation((mediumId) => {
+						if (mediumId === scenario.license.mediumId) {
+							return Promise.reject(new Error('lookup failed'));
+						}
+						return Promise.resolve(secondTool);
+					});
+					return { scenario, secondTool };
+				};
+
+				it('should save the successfully provisioned school external tool', async () => {
+					const { scenario, secondTool } = createPartialFailureScenario();
+
+					await provision(scenario);
+
+					expect(schoolExternalToolService.saveSchoolExternalTool).toHaveBeenCalledTimes(1);
+					expect(schoolExternalToolService.saveSchoolExternalTool).toHaveBeenCalledWith(
+						expect.objectContaining({ toolId: secondTool.id, schoolId: scenario.schoolId })
+					);
+				});
+
+				it('should log a provisioning failure for the failed license', async () => {
+					const { scenario } = createPartialFailureScenario();
+
+					await provision(scenario);
+
+					expect(logger.warning).toHaveBeenCalledWith(expect.any(ExternalToolProvisioningFailedLoggable));
+				});
 			});
 		});
 	});
