+++ apps/server/src/modules/tool/external-tool/repo/mikro-orm/external-tool.entity.ts
@@ -43,15 +43,16 @@ export interface ExternalToolEntityProps {
 
 @Entity({ tableName: 'external-tools' })
 @Unique({ properties: ['config.clientId'], options: { sparse: true } })
+// MongoDB 4 partial indexes support neither $exists: false nor $in, so an unset mediumId is matched via null equality
 @Unique({
 	name: 'externalToolNameUniqueIndex',
 	properties: ['name'],
-	options: { partialFilterExpression: { 'medium.mediumId': { $exists: false } } },
+	options: { partialFilterExpression: { 'medium.mediumId': null } },
 })
 @Unique({
 	name: 'externalToolMediumIdentityUniqueIndex',
 	properties: ['medium.mediumId', 'medium.mediaSourceId'],
-	options: { partialFilterExpression: { 'medium.mediumId': { $type: 'string', $gt: '' } } },
+	options: { partialFilterExpression: { medium: { $exists: true } } },
 })
 export class ExternalToolEntity extends BaseEntityWithTimestamps {
 	@Property()
